Fixed file structure for routes. Added comments explaining functionality of routes.
