+++ public/js/logout.js
@@ -1,6 +1,6 @@
 const logout = async (e) => {
   e.preventDefault();
-  const response = await fetch('logout', {
+  const response = await fetch('api/users/logout', {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
   });
